requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.3
django-allauth==0.43.0
django-bkash-integration==0.3
django-model-utils==4.0.0
django-oauth-toolkit==1.3.3
django-rest-auth==0.9.5
djangorestframework==3.12.2
filelock==3.0.12
hashids==1.3.1
idna==2.10
Markdown==3.3.3
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.1.0
